remove uncessary component
