+++ service/defaultcomponents/defaults.go
@@ -18,37 +18,14 @@ package defaultcomponents
 import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/component/componenterror"
-	"go.opentelemetry.io/collector/exporter/fileexporter"
-	"go.opentelemetry.io/collector/exporter/jaegerexporter"
-	"go.opentelemetry.io/collector/exporter/kafkaexporter"
 	"go.opentelemetry.io/collector/exporter/loggingexporter"
-	"go.opentelemetry.io/collector/exporter/opencensusexporter"
-	"go.opentelemetry.io/collector/exporter/otlpexporter"
-	"go.opentelemetry.io/collector/exporter/prometheusexporter"
 	"go.opentelemetry.io/collector/exporter/prometheusremotewriteexporter"
-	"go.opentelemetry.io/collector/exporter/zipkinexporter"
 	"go.opentelemetry.io/collector/extension/fluentbitextension"
 	"go.opentelemetry.io/collector/extension/healthcheckextension"
 	"go.opentelemetry.io/collector/extension/pprofextension"
 	"go.opentelemetry.io/collector/extension/zpagesextension"
-	"go.opentelemetry.io/collector/processor/attributesprocessor"
 	"go.opentelemetry.io/collector/processor/batchprocessor"
-	"go.opentelemetry.io/collector/processor/filterprocessor"
-	"go.opentelemetry.io/collector/processor/groupbytraceprocessor"
-	"go.opentelemetry.io/collector/processor/memorylimiter"
-	"go.opentelemetry.io/collector/processor/queuedprocessor"
-	"go.opentelemetry.io/collector/processor/resourceprocessor"
-	"go.opentelemetry.io/collector/processor/samplingprocessor/probabilisticsamplerprocessor"
-	"go.opentelemetry.io/collector/processor/samplingprocessor/tailsamplingprocessor"
-	"go.opentelemetry.io/collector/processor/spanprocessor"
-	"go.opentelemetry.io/collector/receiver/fluentforwardreceiver"
-	"go.opentelemetry.io/collector/receiver/hostmetricsreceiver"
-	"go.opentelemetry.io/collector/receiver/jaegerreceiver"
-	"go.opentelemetry.io/collector/receiver/kafkareceiver"
-	"go.opentelemetry.io/collector/receiver/opencensusreceiver"
 	"go.opentelemetry.io/collector/receiver/otlpreceiver"
-	"go.opentelemetry.io/collector/receiver/prometheusreceiver"
-	"go.opentelemetry.io/collector/receiver/zipkinreceiver"
 )
 
 // Components returns the default set of components used by the
@@ -70,45 +47,22 @@ func Components() (
 	}
 
 	receivers, err := component.MakeReceiverFactoryMap(
-		jaegerreceiver.NewFactory(),
-		fluentforwardreceiver.NewFactory(),
-		zipkinreceiver.NewFactory(),
-		prometheusreceiver.NewFactory(),
-		opencensusreceiver.NewFactory(),
 		otlpreceiver.NewFactory(),
-		hostmetricsreceiver.NewFactory(),
-		kafkareceiver.NewFactory(),
 	)
 	if err != nil {
 		errs = append(errs, err)
 	}
 
 	exporters, err := component.MakeExporterFactoryMap(
-		opencensusexporter.NewFactory(),
-		prometheusexporter.NewFactory(),
 		prometheusremotewriteexporter.NewFactory(),
 		loggingexporter.NewFactory(),
-		zipkinexporter.NewFactory(),
-		jaegerexporter.NewFactory(),
-		fileexporter.NewFactory(),
-		otlpexporter.NewFactory(),
-		kafkaexporter.NewFactory(),
 	)
 	if err != nil {
 		errs = append(errs, err)
 	}
 
 	processors, err := component.MakeProcessorFactoryMap(
-		attributesprocessor.NewFactory(),
-		resourceprocessor.NewFactory(),
-		queuedprocessor.NewFactory(),
 		batchprocessor.NewFactory(),
-		memorylimiter.NewFactory(),
-		tailsamplingprocessor.NewFactory(),
-		probabilisticsamplerprocessor.NewFactory(),
-		spanprocessor.NewFactory(),
-		filterprocessor.NewFactory(),
-		groupbytraceprocessor.NewFactory(),
 	)
 	if err != nil {
 		errs = append(errs, err)
